requirements.txt
pandas==1.5.3
spacy==3.5.0
python-dateutil==2.8.2